haproxy: Experimental: add haproxy, and configure it to act as a frontend for the webui SSL connections
httpd/mssl SSL performance is rather poor, in large part due to httpd being single-threaded,
and only supporting basic HTTP.  haproxy supports HTTP2.0 and multiple concurrent
sessions, greatly improving https performance with the webui, and resolving the
page load stalls occuring with Chrome.

This is still experimental, and there are a few additionnal adjustments
required, for instance httpd must be started even if the user choses
https only, as it's used as the backend.

This commit adds haproxy, and takes care of starting haproxy instead of
httpds.  It also updates the watchdog to monitor haproxy instead of
httpds.

Further potential changes could be to remove mssl, and disable the https
code in httpd to reduce a bit on the footprint.