nvram: implement jffs-based storage for dhcp_staticlist on HND
On the HND platform, Broadcom enforces nvram size limits within
closed source code (1000 chars for know nvars, 255 for
unknown vars).  Unfortunately, this means that we can no
longer enlarge these limits on that platform.

For dhcp_staticlist, adding the hostname field can severely
increaes the amount of data stored in that nvram, making it
quicky reach the 1000 characters limit.  And since closed
source components (like networkmap) need to be able to
access that list, we cannot simply spread it accross
multiple nvram entries (like OpenVPN's clientlist).

As a solution, the /jffs/nvramext/ folder is created on the
HND platform, and any nvram starting with the nvext_ prefix will
get stored there instead of in nvram.  The nvram_ext_get() EJ
function was added to retrieve the content of a variable
stored there (as a replacement for nvram_get() in these cases).

(Potential improvement: we could have nvram_get() automatically
access the jffs content on the nvext_ prefix)

For now, only dhcp_staticlist gets stored there.  For backward
compatibility with networkmap (and others), we still store a copy
of static reservations in there, but without the hostname, so the
content will be shorter, and still be usable by networkmap.  The
nvext content will have the extra hostname field.

This is only implemented for HND, as it has one annoying drawback:
data stored in that folder will not be included in a Settings
backup (you'll need a separate JFFS backup for that).